social laws, domains, jsons

+++ experimentation/ProblemGenerator.py
@@ -1076,8 +1076,8 @@ class ExpeditionGenerator(NumericProblemGenerator):
             sl.add_effect(a.name, 'retrieve_supplies', 'personal_packs', ('w',), 1, '-')
             sl.add_effect(a.name, 'store_supplies', 'personal_packs', ('w',), 1, '+')
             sl.set_initial_value_for_new_fluent(a.name, 'personal_packs', ('wa0', ),
-                                                float(self.instance_data['init_values']['global'][0][2]) / len(
-                                                    self.problem.agents))
+                                                int(float(self.instance_data['init_values']['global'][0][2]) / len(
+                                                    self.problem.agents)))
         self.problem = sl.compile(self.problem).problem
         return self.problem
